+++ setup.py
@@ -4,18 +4,18 @@ from setuptools import setup, find_packages
 
 setup(
     name="tap-rest-api",
+    packages=["tap_rest_api"],
     version="0.9.0",
-    description="Singer.io tap for extracting data from the any generic REST API",
+    description="Singer.io tap for extracting data from any generic REST API",
     author="github.com/aaronsteers",
     url="https://github.com/aaronsteers/tap-rest-api",
     classifiers=["Programming Language :: Python :: 3 :: Only"],
-    py_modules=["tap_rest_api"],
+    # py_modules=["tap_rest_api"],
     install_requires=["singer-python==5.3.3", "requests==2.20.0"],
     entry_points="""
           [console_scripts]
-          tap-rest-api=tap_rest_api:cli
+          tap-rest-api=tap_rest_api.cli:main
       """,
-    packages=["tap_rest_api"],
-    package_data={"catalog": ["catalog/*.json"]},
+    package_data={"": ["tap_rest_api/catalog/*.json"]},
     include_package_data=True,
 )
